refactor: removed unnecessary comments
