+++ scripts/zip.js
@@ -4,10 +4,10 @@ import path from 'path';
 
 async function zipDistFolder() {
   const output = createWriteStream(path.resolve('scrapegraphai_extension.zip'));
-  const archive = archiver('zip', { zlib: { level: 9 } }); // Livello massimo di compressione
+  const archive = archiver('zip', { zlib: { level: 9 } });
 
   output.on('close', () => {
-    console.log(`dist.zip creato con successo. Dimensione: ${archive.pointer()} byte`);
+    console.log(`Success! zip file created. Size: ${archive.pointer()} byte`);
   });
 
   archive.on('error', (err) => {
@@ -16,10 +16,9 @@ async function zipDistFolder() {
 
   archive.pipe(output);
 
-  // Aggiungi la cartella dist al file zip
   archive.directory(path.resolve('dist'), false);
 
   await archive.finalize();
 }
 
-zipDistFolder().catch((err) => console.error('Errore durante la creazione dello zip:', err));
+zipDistFolder().catch((err) => console.error('Error during zip file creation:', err));
