--- src/components/alerts/ErrorAlert.js
@@ -1,32 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-const ErrorAlert = ({ message }) => {
-  return (
-    <div className="flex items-center p-4 mb-4 text-sm text-red-800 rounded-lg bg-red-50" role="alert">
-      <svg
-        className="w-6 h-6 me-3"
-        aria-hidden="true"
-        xmlns="http://www.w3.org/2000/svg"
-        width="24"
-        height="24"
-        fill="none"
-        viewBox="0 0 24 24">
-        <path
-          stroke="currentColor"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth="2"
-          d="M12 13V8m0 8h.01M21 12a9 9 0 1 1-18 0 9 9 0 0 1 18 0Z"
-        />
-      </svg>
-      {message}
-    </div>
-  );
-};
-
-ErrorAlert.propTypes = {
-  message: PropTypes.string.isRequired,
-};
-
-export default ErrorAlert;
