--- src/components/common/Header.js
@@ -1,54 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import Logo from '../../assets/scrapegraphai_logo.svg';
-
-const Header = () => {
-  return (
-    <header>
-      <nav className="bg-white p-4">
-        <div className="flex flex-wrap justify-between items-center mx-auto">
-          <a className="flex items-center">
-            <img src={Logo} className="mr-3 h-10" alt="ScrapeGraphAI-Logo" />
-            <span className="self-center text-lg font-semibold whitespace-nowrap">ScrapeGraphAI</span>
-          </a>
-          <div className="flex items-center lg:order-2">
-            <Link to="/settings">
-              <button
-                data-collapse-toggle="mobile-menu-2"
-                type="button"
-                className="inline-flex items-center p-2 ml-1 text-sm text-gray-500 rounded-lg lg:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200"
-                aria-controls="mobile-menu-2"
-                aria-expanded="false">
-                <svg
-                  className="w-6 h-6 text-gray-800"
-                  aria-hidden="true"
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="24"
-                  height="24"
-                  fill="none"
-                  viewBox="0 0 24 24">
-                  <path
-                    stroke="currentColor"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    d="M21 13v-2a1 1 0 0 0-1-1h-.757l-.707-1.707.535-.536a1 1 0 0 0 0-1.414l-1.414-1.414a1 1 0 0 0-1.414 0l-.536.535L14 4.757V4a1 1 0 0 0-1-1h-2a1 1 0 0 0-1 1v.757l-1.707.707-.536-.535a1 1 0 0 0-1.414 0L4.929 6.343a1 1 0 0 0 0 1.414l.536.536L4.757 10H4a1 1 0 0 0-1 1v2a1 1 0 0 0 1 1h.757l.707 1.707-.535.536a1 1 0 0 0 0 1.414l1.414 1.414a1 1 0 0 0 1.414 0l.536-.535 1.707.707V20a1 1 0 0 0 1 1h2a1 1 0 0 0 1-1v-.757l1.707-.708.536.536a1 1 0 0 0 1.414 0l1.414-1.414a1 1 0 0 0 0-1.414l-.535-.536.707-1.707H20a1 1 0 0 0 1-1Z"
-                  />
-                  <path
-                    stroke="currentColor"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth="2"
-                    d="M12 15a3 3 0 1 0 0-6 3 3 0 0 0 0 6Z"
-                  />
-                </svg>
-              </button>
-            </Link>
-          </div>
-        </div>
-      </nav>
-    </header>
-  );
-};
-
-export default Header;
